backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
boto3>=1.34.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
psutil>=5.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
email_validator>=2.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
